requirements.txt
dill==0.3.5.1
dnspython==2.2.1
evidently==0.1.58.dev0
fastapi==0.78.0
httptools==0.5.0
imblearn==0.0
mypy-boto3-s3==1.24.76
pip-chill==1.0.1
pymongo==4.2.0
python-dotenv==0.21.0
types-s3transfer==0.6.0.post4
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
wincertstore==0.2
xgboost==1.6.2
pandas
PyYAML
apache-airflow

# Flask Framework
Flask==1.0.4

# Flask Packages
Flask-Login==0.6.2.
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.15
Flask-User==1.0.1.5
Jinja2==3.0.0
MarkupSafe==2.1.1
Flask-JWT-Extended == 4.4.4
Flask-Appbuilder == 4.1.4
-e .
